Add a rainbowgram variation to CQT plugin.
This uses the rainbowgram (amplitude and instantaneous frequency) idea, adapted for the real outputs used in Sonic Visualizer.

This variation is designed to be used with a rainbow color palette divided into intervals, where each interval goes from H=0° to 360°. The phase information is encoded as hue, with the magnitude determining the overall intensity.

Sources:
- https://librosa.org/doc/main/auto_examples/plot_rainbowgram.html